refactor(ai): relocate label-halo helpers to colors.py to stay within module budget
The luminance-based halo helpers added for readable AI labels pushed
chat_actions.py to 422 lines, over the 400-line decomposed-module budget
(test_layering.py). They are pure color utilities, so move them to
processing/ai/colors.py (already imported by chat_actions) and expose
label_halo_color as the public entry point alongside is_css_colorish.
chat_actions.py returns to 378 lines; no cap bump needed, guard stays honest.
